[mod] 단어 기능 수정

+++ src/main/kotlin/com/hunmin/domain/service/WordService.kt
@@ -4,12 +4,11 @@ import com.hunmin.domain.dto.word.WordRequestDTO
 import com.hunmin.domain.dto.word.WordResponseDTO
 import com.hunmin.domain.entity.Member
 import com.hunmin.domain.entity.MemberRole
+import com.hunmin.domain.entity.Word
 import com.hunmin.domain.exception.AdminException
-import com.hunmin.domain.exception.MemberException
 import com.hunmin.domain.exception.WordException
 import com.hunmin.domain.repository.MemberRepository
 import com.hunmin.domain.repository.WordRepository
-import org.hibernate.query.sqm.tree.SqmNode.log
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
@@ -21,129 +20,28 @@ class WordService(
     private val wordRepository: WordRepository,
     private val memberRepository: MemberRepository
 ) {
-//    // 관리자 확인
-//    private fun checkAdmin(memberId: Long): Member {
-//        return memberRepository.findById(memberId).orElseThrow {
-//            throw WordException.WORD_FORBIDDEN.toException()
-//        }.also { member ->
-//            if (member.memberRole != MemberRole.ADMIN) {
-//                throw WordException.WORD_FORBIDDEN.toException()
-//            }
-//        }
-//    }
-//
-//    // 단어 등록
-//    fun createWord(wordRequestDTO: WordRequestDTO, memberId: Long): WordResponseDTO {
-//        val member = checkAdmin(memberId) // 관리자 확인 및 Member 객체 조회
-//
-//        return try {
-//            val word = wordRequestDTO.toEntity(member) // Member 객체 전달
-//            val savedWord = wordRepository.save(word)
-//
-//            WordResponseDTO(savedWord)
-//        } catch (e: Exception) {
-//            throw WordException.WORD_NOT_CREATED.toException()
-//        }
-//    }
-    // 관리자 확인
-    private fun checkAdmin(memberId: Long) {
-        val member = memberRepository.findById(memberId).orElseThrow()
-        if (member.memberRole != MemberRole.ADMIN) {
-            throw WordException.WORD_FORBIDDEN.toException()
+    // 회원 확인
+    private fun checkMember(memberId: Long): Member {
+        return memberRepository.findById(memberId).orElseThrow {
+            throw AdminException.MEMBER_NOT_FOUND.get()
         }
     }
 
-//    // 단어 등록
-//    fun createWord(wordRequestDTO: WordRequestDTO, memberId: Long): WordResponseDTO {
-//        checkAdmin(memberId)
-//
-//        return try {
-//            val word = wordRequestDTO.toEntity()
-//            val savedWord = wordRepository.save(word)
-//
-//            WordResponseDTO(savedWord)
-//        } catch (e: Exception) {
-//            throw WordException.WORD_NOT_CREATED.toException()
-//        }
-//    }
-
-
-//    // 멤버 확인
-//    private fun getMember(username: String): Member {
-//        return memberRepository.findByEmail(username) ?: throw MemberException.NOT_FOUND.get()
-//    }
-//
-//    // 단어 등록
-//    fun createWord(wordRequestDTO: WordRequestDTO, username: String): WordResponseDTO {
-//        val member = getMember(username)
-//
-//        if (member.memberRole != MemberRole.ADMIN) {
-//            throw WordException.WORD_FORBIDDEN.toException()
-//        }
-//
-//        return try {
-//            val word = wordRequestDTO.toEntity(member)
-//            val savedWord = wordRepository.save(word)
-//            WordResponseDTO(savedWord)
-//        } catch (e: Exception) {
-//            throw WordException.WORD_NOT_CREATED.toException()
-//        }
-//    }
-
-//    // 단어 수정
-//    fun updateWord(wordRequestDTO: WordRequestDTO, memberId: Long): WordResponseDTO {
-//        checkAdmin(memberId)
-//
-//        val word = wordRepository.findByTitleAndLang(wordRequestDTO.title, wordRequestDTO.lang)
-//            .orElseThrow { WordException.WORD_NOT_FOUND.toException() }
-//
-//        return try {
-//            word.changeTitle(wordRequestDTO.title)
-//            word.changeLang(wordRequestDTO.lang)
-//            word.changeTranslation(wordRequestDTO.translation)
-//            word.changeDefinition(wordRequestDTO.definition)
-//
-//            val updatedWord = wordRepository.save(word)
-//
-//            WordResponseDTO(updatedWord)
-//        } catch (e: Exception) {
-//            throw WordException.WORD_NOT_UPDATED.toException()
-//        }
-//    }
-//
-//    // 단어 삭제
-//    fun deleteWord(title: String, lang: String, memberId: Long) {
-//        checkAdmin(memberId)
-//
-//        val word = wordRepository.findByTitleAndLang(title, lang)
-//            .orElseThrow { WordException.WORD_NOT_FOUND.toException() }
-//
-//        try {
-//            wordRepository.delete(word)
-//        } catch (e: Exception) {
-//            throw WordException.WORD_NOT_DELETED.toException()
-//        }
-//    }
-//
-    // 단어 조회
-    fun getWord(title: String, lang: String): WordResponseDTO {
-        val word = wordRepository.findByTitleAndLang(title, lang)
-            .orElseThrow { WordException.WORD_NOT_FOUND.toException() }
-
-        return WordResponseDTO(word)
+    // 관리자 확인
+    private fun checkAdmin(memberId: Long): Member {
+        val member: Member = checkMember(memberId) // 여기서 멤버 확인
+        if (member.memberRole != MemberRole.ADMIN) {
+            throw WordException.WORD_FORBIDDEN.toException()
+        }
+        return member
     }
 
-    // 단어 전체 조회
-    fun getAllWords(lang: String, pageable: Pageable): Page<WordResponseDTO> {
-        val wordsPage = wordRepository.findByLang(lang, pageable)
-        return wordsPage.map { WordResponseDTO(it) }
-    }
-
-    // -------------------------------------------------------------------------------------------------
-    // 테스트 등록
+    // 단어 등록
     fun testCreate(wordRequestDTO: WordRequestDTO): WordResponseDTO {
+        val member = checkAdmin(wordRequestDTO.memberId ?: throw AdminException.MEMBER_NOT_FOUND.get())
+
         return try {
-            val word = wordRequestDTO.testEntity() // testEntity로 임시 생성
+            val word: Word = wordRequestDTO.toEntity(member)
             val savedWord = wordRepository.save(word)
             WordResponseDTO(savedWord)
         } catch (e: Exception) {
@@ -151,9 +49,11 @@ class WordService(
         }
     }
 
-    // 테스트 수정
+    // 단어 수정
     fun testUpdate(wordRequestDTO: WordRequestDTO): WordResponseDTO {
-        val word = wordRepository.findByTitleAndLang(wordRequestDTO.title, wordRequestDTO.lang)
+        checkAdmin(wordRequestDTO.memberId ?: throw AdminException.MEMBER_NOT_FOUND.get())
+
+        val word = wordRepository.findByTitleAndLang(wordRequestDTO.originalTitle, wordRequestDTO.originalLang)
             .orElseThrow { WordException.WORD_NOT_FOUND.toException() }
 
         return try {
@@ -170,15 +70,35 @@ class WordService(
         }
     }
 
-    // 테스트 삭제
-    fun testDelete(title: String, lang: String) {
+    // 단어 삭제
+    fun testDelete(title: String, lang: String, memberId: Long) {
+        checkAdmin(memberId)
+
         val word = wordRepository.findByTitleAndLang(title, lang)
             .orElseThrow { WordException.WORD_NOT_FOUND.toException() }
 
         try {
-            wordRepository.delete(word)  // 단어 삭제
+            wordRepository.delete(word)
         } catch (e: Exception) {
             throw WordException.WORD_NOT_DELETED.toException()
         }
+    }
+
+    // 단어 조회
+    fun getWord(title: String, lang: String, memberId: Long): WordResponseDTO {
+        checkMember(memberId)
+
+        val word = wordRepository.findByTitleAndLang(title, lang)
+            .orElseThrow { WordException.WORD_NOT_FOUND.toException() }
+
+        return WordResponseDTO(word)
+    }
+
+    // 단어 전체 조회
+    fun getAllWords(lang: String, pageable: Pageable, memberId: Long): Page<WordResponseDTO> {
+        checkMember(memberId)
+
+        val wordsPage = wordRepository.findByLang(lang, pageable)
+        return wordsPage.map { WordResponseDTO(it) }
     }
 }
